+++ src/Mod/TechDraw/Gui/QGVPage.h
@@ -120,6 +120,8 @@ public:
     void setPanCursor();
     void setZoomCursor();
 
+    QPixmap prepareCursorPixmap(const char* iconName, QPoint& hotspot);
+
     void pseudoContextEvent();
 
     void centerOnPage();
@@ -150,8 +152,6 @@ protected:
     void contextMenuEvent(QContextMenuEvent* event) override;
 
     QColor getBackgroundColor();
-
-    QPixmap prepareCursorPixmap(const char* iconName, QPoint& hotspot);
 
     void drawForeground(QPainter* painter, const QRectF& rect) override;
